Fixed bad request

+++ src/main/java/xyz/deseteral/phobot/Chat.java
@@ -4,14 +4,14 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class Chat {
-    private int id;
+    private long id;
 
     @JsonCreator
-    public Chat(@JsonProperty("id") int id) {
+    public Chat(@JsonProperty("id") long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 }
